+++ CHANGELOG.md
@@ -15,7 +15,8 @@ All notable changes to Scrawler will be documented in this file.
 ### Changed
 - Underlying Guzzle instance will _always_ depend on cURL now. This is done to
   ensure that widest set of features is available for handling HTTP requests.
-- `DefaultConfigurationProvider` sets timeouts for Guzzle now  
+- Scrawler will now explicitly emit a warning for content types other than XML or (X)HTML
+- `DefaultConfigurationProvider` sets timeouts for Guzzle now
 - `JSON_UNESCAPED_UNICODE` option is now used by default when using JSON file
   result writer
 - Only HTTP and HTTPS protocols are now explicitly allowed